backend/requirements.txt
fastapi==0.116.1
uvicorn==0.34.0
psycopg2-binary==2.9.10
bcrypt==4.2.1
python-multipart==0.0.17
pydantic==2.10.4
